+++ system/src/main/java/io/github/sds100/keymapper/system/permissions/Permission.kt
@@ -18,4 +18,5 @@ enum class Permission {
     FIND_NEARBY_DEVICES,
     POST_NOTIFICATIONS,
     READ_LOGS,
+    ACCESS_LOCAL_NETWORK,
 }